SC,TSO,PSO: Use stable symbolic memory addresses
Symbolic memory addresses are represented by a new type SymAddr, which
in turn is a SymMBlock and an offset. A SymMBlock corresponds to a
contigous area of memory that the program has access to, it can be
either Global(n), which is the n'th global variable, Heap(t, n), which
is the n'th heap allocation by thread t, or Stack(t, n).