nidhuggc: Allow program arguments
nidhuggc is changed from treating the last argument as the input to
treating the first non-option argument as the input. This does not
change the behaviour for any previously valid command lines as you could
never pass more than one non-option argument to nidhuggc.

Since we're letting nidhugg make the distinction between arguments to
itself and program arguments, there cannot be a difference in how
nidhugg and nidhuggc will parse these new additions.