TSOTraceBuilder: Fix LockFail subsumption bug
LockFail is a "virtual" event, in that it is introduced by the
extensions to Optimal-DPOR for blocking in the race detection phase,
rather than being an event that appears in the execution. As such, it
does not contribute to the vector clock and thus should not be
considered as subsuming any other race.