Adds the ability to run the Reads-From algorithm under SC in parallel.
It operates with a threadpool of worker-threads that will explore leaves
with priority of deepest node first to optimize memory use.

To run RFSC in parallel, introduce the flag "-n-threads=N" with N equal
to the threads to be executed.
The threadpool operate in a multi-producer single-consumer which
operates best with n-threads > 2.

Strict changes That might affect execution

 * Previously reparse() was called after every 1000 executions. We were
   told that this was due to a bug that should be resolved, this
   hindered our concurrent runners and didn't seem to be needed anymore.
 * estimate-progress relies on the size of a TraceBuilder-prefix, as
   these run concurrently we did not solve this ability and do not
   update the value.
 * Timing.cpp relies on exclusive behaviour, This have been deactivated
   for the exploration to run.

Co-authored-by: kankava <kankava@users.noreply.github.com>