Make LLVMUtils::getPthreadTType smarter
Now, it looks up if we have definitions of "pthread_join" or
"pthread_self", and takes the definition from there.

This is needed when the IR module comes from another system where the
pthread type is different. Or, as in what motivated this change, we have
hardcoded IR in unit tests and want those tests to work everywhere.